requirements.txt
azure-identity==1.13.0b4
Flask==2.3.2
openai==0.27.7
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
python-dotenv==1.0.0
azure-cosmosdb-table==1.0.6
presidio-analyzer==2.2.32
presidio-anonymizer==2.2.32
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
pandas